refactor: move use-rpc-handler to /shared
being used by both frontend & backend
